Put the game provider in its own header.
